% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{accurate_sum}
\alias{accurate_sum}
\title{Accurate sum of numbers}
\usage{
accurate_sum(v)
}
\description{
Sums every time only two smallest numbers.
This allows calculate the sum of numbers accurate enough.
}
