% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{get_frequency}
\alias{get_frequency}
\title{Frequency of vertecies.}
\usage{
get_frequency(graph, mat, inds = seq_len(nrow(mat)))
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{mat}{Matrix object.}

\item{inds}{Index numbers of rows involved in the calculation.}
}
\value{
Named vector of frequency.
}
\description{
Calculates the frequency of occurences of vertices in matrix object.
}
\seealso{
\code{\link{get_prob}}
}
