% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{get_not_prob}
\alias{get_not_prob}
\title{Vertex probability using MCMC.}
\usage{
get_not_prob(mcmcObject)
}
\arguments{
\item{mcmcObj}{Object of type MCMC.}
}
\value{
list.
}
\description{
Calculates the probability of vertices using its frequency of occurence in matrix object.
}
\seealso{
\code{\link{get_frequency}}
}
