% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{get_prob}
\alias{get_prob}
\title{Vertex probability using MCMC.}
\usage{
get_prob(mcmcObject, inds = seq_len(nrow(mcmcObject$mat)))
}
\arguments{
\item{inds}{Index numbers of rows involved in the calculation.}
}
\value{
Named vector of probabilities.
}
\description{
Calculates the probability of vertices using its frequency of occurence in matrix object.
}
\seealso{
\code{\link{get_frequency}}
}
