% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{get_reverse_rank}
\alias{get_reverse_rank}
\title{Ranking vertices.}
\usage{
get_reverse_rank(graph, q)
}
\arguments{
\item{graph}{An object of type \code{igraph}.}

\item{q}{score.}
}
\value{
Named rank vector.
}
\description{
Ranking vertices according to the score.
}
\seealso{
\code{\link{get_rank}}
}
