% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc}
\alias{mcmc}
\title{MCMC class constructor}
\usage{
mcmc(mat, fixed_order)
}
\arguments{
\item{mat}{A boolean matrix object where every row defines a subgraph.}

\item{fixed_order}{Logical scalar. whether a matrix was obtained from fixed order subgraphs.}
}
\value{
Object of class MCMC.
}
\description{
MCMC data structure contains information about subgraph.
}
