% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Set of connected subgraphs from uniform distribution.}
\usage{
mcmc_sample(graph, module_size = NULL, times = NULL, previous_mcmc = NULL,
  niter, exp_lh, fixed_size = FALSE)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{times}{Number of subgraphs.}

\item{previous_mcmc}{Object of type MCMC.}

\item{niter}{Number of iterations.}
}
\value{
Object of type MCMC.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\examples{
graph <- barabasi.game(20, 1.2, 2, directed = F)
V(graph)$name <- letters[1:20]
V(graph)$likelihood <- 1
x <- mcmc_sample(graph = graph, module_size = 5, iter = 10, times = 4)
matrix(x$name[x$mat], nrow(x$mat))
}
\seealso{
\code{\link{sample_subgraph}, \link{mcmc_onelong}}
}
