% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_subgraph}
\alias{mcmc_subgraph}
\title{Connected subgraph from uniform distribution.}
\usage{
mcmc_subgraph(graph, module_size, iter)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{iter}{Number of iterations.}
}
\value{
Vector of vertex names of connected subgraph.
}
\description{
Generates a connected subgraph using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{mcmc_sample}, \link{mcmc_onelong}}
}
