% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{mcmcRanking}
\alias{mcmcRanking}
\alias{mcmcRanking-package}
\title{The mcmcRanking package}
\description{
Tool for estimate probabilities of vertices being in active subgraph using
its likelihoods.
}
\author{
Javlon Isomurodov <isomurodov@rain.ifmo.ru>
}
