% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.R
\name{probabilistic_rank}
\alias{probabilistic_rank}
\title{Probabilistic ranking.}
\usage{
probabilistic_rank(graph, q)
}
\arguments{
\item{graph}{An object of type \code{igraph}.}

\item{q}{Vector of probabilities of vertices to not belong to active module.}
}
\value{
ranking.
}
\description{
Heuristic ranking maximizing the area under the ROC curve.
}
\examples{
library(igraph)
data(exampleGraph)
q <- V(exampleGraph)$q
names(q) <- V(exampleGraph)$name
rank <- probabilistic_rank(exampleGraph, q)
head(rank)
}
\seealso{
\code{\link{mcmc_sample}}
}
