% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{real_prob}
\alias{real_prob}
\title{Vertex probability.}
\usage{
real_prob(graph)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}
}
\value{
Named vector of probabilites.
}
\description{
Accurate estimate of vertex probability using all connected subgraphs.
}
\details{
Time complexity of method is \strong{exponential}. Use it only for the graphs of size less than 30.
}
\seealso{
\code{\link{get_prob}}
}
