% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repetition_depth}
\alias{repetition_depth}
\title{Number of runs of iterative likelihood increase of vertices.}
\usage{
repetition_depth(x)
}
\arguments{
\item{x}{According to this value repetition depth calculates.}
}
\value{
A number of runs.
}
\description{
Running MCMC with likelihood weights on vertices can cause to get stuck in locally
good vertices, thereby finding locally significant solutions.
So, we will iteratively increase vertex weights and this function calculates number of necessary iterations.
}
\examples{
repetition_depth(16)
repetition_depth(81)
}
