% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{sample_llh}
\alias{sample_llh}
\title{Sample log-likelihood values during one run.}
\usage{
sample_llh(graph, subgraph_order, niter, exp_lh = 1,
  fixed_order = FALSE)
}
\arguments{
\item{graph}{The original graph of class \code{igraph}.}

\item{subgraph_order}{The order of subgraph.}

\item{niter}{Number of iterations the MCMC method will run.}

\item{exp_lh}{The power to which the likelihood values will be raised}

\item{fixed_order}{Logical scalar. Whether to fix the subgraph order.}
}
\value{
A named vector of likelihoods where names are number of iteration.
}
\description{
Generates log-likelihood values during one Markov chain Monte Carlo (MCMC)
run for analyzing behavior of subgraph.
}
\details{
\code{graph} must contain vertex attribute \emph{likelihood}.
}
\examples{
data(exampleGraph)
llhs <- sample_llh(exampleGraph, 0, 1e5)
tail(llhs)
}
