% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{sample_subgraph}
\alias{sample_subgraph}
\title{Sample a connected subgraph from an uniform distribution.}
\usage{
sample_subgraph(graph, subgraph_order, niter)
}
\arguments{
\item{graph}{The original graph of class \code{igraph}.}

\item{subgraph_order}{The order of subgraph.}

\item{niter}{Number of iterations the MCMC method will run.}
}
\value{
Character vector.
}
\description{
Generates a connected subgraph from an uniform distribution using Markov
chain Monte Carlo (MCMC) method.
}
\examples{
data(exampleGraph)
sample_subgraph(exampleGraph, 10, 1e4)
}
\seealso{
\code{\link{mcmc_sample}}
}
