% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_likelihood}
\alias{set_likelihood}
\title{Set likelihood.}
\usage{
set_likelihood(graph, fdr)
}
\arguments{
\item{graph}{An \code{igraph} graph.}

\item{fdr}{Numeric constant, from the false discovery rate a p-value threshold is calculated.}
}
\value{
An \code{igraph} graph with \emph{likelihood} vertex attribute.
}
\description{
Set \emph{likelihood} vertex attribute to an \code{igraph} graph using pval vertex attribute.
}
\examples{
data(exampleGraph)
set_likelihood(exampleGraph, 1e-7)
}
