% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_frequency}
\alias{get_frequency}
\title{Frequency of vertecies.}
\usage{
get_frequency(mcmcObj, inds = seq_len(nrow(mcmcObj$mat)))
}
\arguments{
\item{mcmcObj}{Object of type MCMC.}

\item{inds}{Index numbers of rows involved in the calculation.}
}
\value{
Named vector of frequency.
}
\description{
Calculates the frequency of occurences of vertices in matrix object.
}
\seealso{
\code{\link{get_prob}}
}
