% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_frequency}
\alias{get_frequency}
\title{Frequency of vertices.}
\usage{
get_frequency(mcmcObj, inds = seq_len(nrow(mcmcObj$mat)))
}
\arguments{
\item{mcmcObj}{Object of class MCMC.}

\item{inds}{Index numbers of rows involved in the calculation.}
}
\value{
A named vector of frequency.
}
\description{
Calculates the frequency of occurrences of vertices in matrix object.
}
\examples{
data(exampleGraph)
x <- mcmc_sample(exampleGraph, subgraph_order = 0, times = 1e3, niter = 100)
freq <- get_frequency(x)
tail(sort(freq))
}
\seealso{
\code{\link{mcmc_sample}, \link{mcmc_onelong}}
}
