% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{get_prob}
\alias{get_prob}
\title{Vertex probability using MCMC.}
\usage{
get_prob(graph, mat, inds = seq_len(nrow(mat)))
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{mat}{Matrix object.}

\item{inds}{Index numbers of rows involved in the calculation.}
}
\value{
Named vector of probabilities.
}
\description{
Calculates the probability of vertices using its frequency of occurence in matrix object.
}
\seealso{
\code{\link{get_frequency}}
}
