% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{get_rank}
\alias{get_rank}
\title{Ranking vertices.}
\usage{
get_rank(graph, q, size)
}
\arguments{
\item{graph}{An object of type \code{igraph}.}

\item{q}{score.}

\item{size}{The initial size.}
}
\value{
Named rank vector.
}
\description{
Ranking vertices according to the score.
}
\seealso{
\code{\link{get_reverse_rank}}
}
