% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc}
\alias{mcmc}
\title{MCMC class constructor}
\usage{
mcmc(mat, name)
}
\arguments{
\item{mat}{A boolean matrix object where every row defines a subgraph.}

\item{name}{A character vector contains names of graph vertices.}
}
\value{
Object of class MCMC.
}
\description{
MCMC data structure contains information about subgraph.
}
