% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_onelong}
\alias{mcmc_onelong}
\title{Sample connected subgraphs by likelihood value during one long MCMC run.}
\usage{
mcmc_onelong(graph, subgraph_order, start, niter)
}
\arguments{
\item{graph}{The original graph of class \code{igraph}.}

\item{subgraph_order}{The order of subgraph.}

\item{start}{Starting with this iteration, we write down all states of the
Markov process.}

\item{niter}{Number of iterations the MCMC method will run.}
}
\value{
Object of class MCMC.
}
\description{
Generates set of subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\details{
\code{graph} must contain vertex attribute \emph{likelihood}.
}
\examples{
data(exampleGraph)
x <- mcmc_onelong(exampleGraph, 50, 1e4, 2e4)
freq <- get_frequency(x)
tail(sort(freq))
}
\seealso{
\code{\link{mcmc_onelong_frequency}, \link{mcmc_sample},
  \link{get_frequency}}
}
