% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_onelong}
\alias{mcmc_onelong}
\title{Sampling set of connected subgraphs by likelihood value.}
\usage{
mcmc_onelong(graph, module_size, start, end)
}
\arguments{
\item{graph}{An \code{igraph} graph with \code{lieklihood} vertex attribute.}

\item{module_size}{The size of subgraph.}

\item{start}{Starting with this iteration, we write down all states of the Markov process.}

\item{end}{Number of iterations.}
}
\value{
Object of class MCMC.
}
\description{
Generates set of subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{sample_subgraph}, \link{mcmc_sample}}
}
