% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_onelong}
\alias{mcmc_onelong}
\title{Set of connected subgraphs.}
\usage{
mcmc_onelong(graph, module_size, start, end, fun = function(x) x)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{start}{Starting with this iteration, we write down all states of the Markov process.}

\item{end}{Number of iterations.}
}
\value{
Object of type MCMC.
}
\description{
Generates set of subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{mcmc_subgraph}, \link{mcmc_sample}}
}
