% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_onelong_frequency}
\alias{mcmc_onelong_frequency}
\title{Frequency of vertices.}
\usage{
mcmc_onelong_frequency(graph, module_size, start, end)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{start}{Starting with this iteration, we write down all states of the Markov process.}

\item{end}{Number of iterations.}
}
\value{
Named frequency vector.
}
\description{
The frequency of occurrence of vertices in one long run of method Markov chain Monte Carlo (MCMC).
}
\seealso{
\code{\link{mcmc_onelong}, \link{sample_subgraph}, \link{mcmc_sample}}
}
