% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_onelong_frequency}
\alias{mcmc_onelong_frequency}
\title{The frequency of occurrence of vertices in one long MCMC run.}
\usage{
mcmc_onelong_frequency(graph, subgraph_order, start, niter,
  fixed_size = FALSE)
}
\arguments{
\item{graph}{The original graph of class \code{igraph}.}

\item{subgraph_order}{The order of subgraph.}

\item{start}{Starting with this iteration, we write down all states of the
Markov process.}

\item{niter}{Number of iterations the MCMC method will run.}

\item{fixed_size}{Logical scalar. Whether to fix the module order.}
}
\value{
A named frequency vector.
}
\description{
The frequency of occurrence of vertices in one long run of method Markov
chain Monte Carlo (MCMC).
}
\examples{
data(exampleGraph)
freq <- mcmc_onelong_frequency(exampleGraph, 50, 1e4, 2e4)
tail(sort(freq), 60)
}
\seealso{
\code{\link{mcmc_onelong}, \link{sample_subgraph},
  \link{mcmc_sample}}
}
