% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_onelong_frequency}
\alias{mcmc_onelong_frequency}
\title{Frequency of vertices using one long run in MCMC.}
\usage{
mcmc_onelong_frequency(graph, module_size, start, end)
}
\arguments{
\item{graph}{An \code{igraph} graph with \code{lieklihood} vertex attribute.}

\item{module_size}{The size of subgraph.}

\item{start}{Starting with this iteration, we write down all states of the Markov process.}

\item{end}{Number of iterations.}
}
\value{
A named frequency vector.
}
\description{
The frequency of occurrence of vertices in one long run of method Markov chain Monte Carlo (MCMC).
}
\examples{
data(exampleGraph)
freq <- mcmc_onelong_frequency(exampleGraph, 50, 1e4, 2e4)
tail(sort(freq), 60)
}
\seealso{
\code{\link{mcmc_onelong}, \link{sample_subgraph}, \link{mcmc_sample}}
}
