% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Set of connected subgraphs from uniform distribution.}
\usage{
mcmc_sample(graph, module_size = NULL, times = NULL, previous_mcmc = NULL,
  iter, fixed_size = TRUE, fun = function(x) x, nproc = 0,
  BPPARAM = NULL, granularity = 10)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{times}{Number of subgraphs.}

\item{previous_mcmc}{Object of type MCMC.}

\item{iter}{Number of iterations.}
}
\value{
Object of type MCMC.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\examples{
graph <- barabasi.game(20, 1.2, 2, directed = F)
V(graph)$name <- letters[1:20]
V(graph)$likelihood <- 1
x <- mcmc_sample(graph = graph, module_size = 5, iter = 10, times = 4)
matrix(x$name[x$mat], nrow(x$mat))
}
\seealso{
\code{\link{mcmc_subgraph}, \link{mcmc_onelong}}
}
