% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Sample connected subgraphs by likelihood value.}
\usage{
mcmc_sample(graph, subgraph_order, times, niter, previous_mcmc,
  exp_lh = 1)
}
\arguments{
\item{graph}{The original graph of class \code{igraph}.}

\item{subgraph_order}{The order of subgraph.}

\item{times}{A number of sampling subgraphs.}

\item{niter}{Number of iterations the MCMC method will run.}

\item{previous_mcmc}{Object of class MCMC.}

\item{exp_lh}{The power to which the likelihood values will be raised}
}
\value{
Object of class MCMC.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC)
method.
}
\details{
\code{graph} must contain vertex attribute \emph{likelihood}.
}
\examples{
data(exampleGraph)
x <- mcmc_sample(exampleGraph, subgraph_order = 0, times = 1e3, niter = 100)
freq <- get_frequency(x)
tail(sort(freq))
}
\seealso{
\code{\link{sample_llh}, \link{sample_subgraph},
  \link{mcmc_onelong}, \link{get_frequency}}
}
