% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Set of connected subgraphs from uniform distribution.}
\usage{
mcmc_sample(graph, module_size = NULL, start_module = NULL, iter,
  times = 1, fun = function(x) x, nproc = 0, BPPARAM = NULL,
  granularity = 10)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{start_module}{Vector of initial module vertecies names.}

\item{iter}{Number of iterations.}

\item{times}{Number of subgraphs.}
}
\value{
Object of type MCMC.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{mcmc_subgraph}, \link{mcmc_onelong}}
}
