% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Set of connected subgraphs from uniform distribution.}
\usage{
mcmc_sample(graph, module_size, iter, times = 1)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{iter}{Number of iterations.}

\item{times}{Number of subgraphs.}
}
\value{
Matrix, where every row corresponds to one subgraph.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{mcmc_subgraph}, \link{mcmc_onelong}}
}
