% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Set of connected subgraphs from uniform distribution.}
\usage{
mcmc_sample(graph, module_size = NULL, times = NULL, previous_mcmc = NULL,
  niter, exp_lh = 1, fixed_size = FALSE)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{times}{Number of subgraphs.}

\item{previous_mcmc}{Object of type MCMC.}

\item{niter}{Number of iterations.}

\item{exp_lh}{The exponent likelihood values to be raised to.}

\item{fixed_size}{\code{TRUE} if the module size is fixed.}
}
\value{
Object of type MCMC.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{sample_subgraph}, \link{mcmc_onelong}}
}
