% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{mcmc_sample}
\alias{mcmc_sample}
\title{Sampling set of connected subgraphs by likelihood value.}
\usage{
mcmc_sample(graph, module_size, times, niter, previous_mcmc, exp_lh = 1,
  fixed_size = FALSE)
}
\arguments{
\item{graph}{An \code{igraph} graph with \code{lieklihood} vertex attribute.}

\item{module_size}{The size of subgraph.}

\item{times}{A number of subgraphs.}

\item{niter}{Number of iterations.}

\item{previous_mcmc}{Object of class MCMC.}

\item{exp_lh}{The exponent likelihood values to be raised to.}

\item{fixed_size}{\code{TRUE} if the module size is fixed.}
}
\value{
Object of class MCMC.
}
\description{
Generates set of independent subgraphs using Markov chain Monte Carlo (MCMC) method.
}
\examples{
data(exampleGraph)
x <- mcmc_sample(exampleGraph, module_size = 0, times = 1e3, niter = 100)
freq <- get_frequency(x)
tail(sort(freq))
}
\seealso{
\code{\link{sample_llh}, \link{sample_subgraph}, \link{mcmc_onelong}}
}
