% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{probabilistic_rank}
\alias{probabilistic_rank}
\title{Probabilistic ranking.}
\usage{
probabilistic_rank(graph, q)
}
\arguments{
\item{graph}{An object of type \code{igraph}.}

\item{q}{Vector of probabilities of vertices to not belong to active module.}
}
\value{
ranking.
}
\description{
Heuristic ranking maximizing the area under the ROC curve.
}
\seealso{
\code{\link{mcmc_sample}}
}
