% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{sample_llh}
\alias{sample_llh}
\title{Generates log-likelihood values during one MCMC run.}
\usage{
sample_llh(graph, module_size = NULL, niter, exp_lh = 1,
  fixed_size = FALSE)
}
\arguments{
\item{graph}{An object of type \code{igraph} with \code{lieklihood} field in vertices.}

\item{module_size}{The size of subgraph.}

\item{niter}{Number of iterations.}

\item{exp_lh}{The exponent likelihood values to be raised to.}

\item{fixed_size}{\code{TRUE} if the module size is fixed.}
}
\value{
Named vector where names are number of iteration.
}
\description{
Generates log-likelihood values during one MCMC run for analyzing behavior of subgraph.
}
