% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{sample_subgraph}
\alias{sample_subgraph}
\title{Connected subgraph from uniform distribution.}
\usage{
sample_subgraph(graph, module_size, niter)
}
\arguments{
\item{graph}{An \code{igraph} graph with \code{lieklihood} vertex attribute.}

\item{module_size}{The size of subgraph.}

\item{niter}{Number of iterations.}
}
\value{
Vector of vertex names of connected subgraph.
}
\description{
Generates a connected subgraph using Markov chain Monte Carlo (MCMC) method.
}
\seealso{
\code{\link{mcmc_sample}, \link{mcmc_onelong}}
}
