% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_likelihood}
\alias{set_likelihood}
\title{Set likelihood.}
\usage{
set_likelihood(graph, fdr)
}
\arguments{
\item{graph}{An object of type \code{igraph}.}

\item{fdr}{Numeric constant, from the false discovery rate a p-value threshold is calculated.}
}
\value{
igraph object.
}
\description{
Set likelihood attribute to vertices of graph using pval attribute.
}
