% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{set_likelihood}
\alias{set_likelihood}
\title{Set likelihood.}
\usage{
set_likelihood(graph)
}
\arguments{
\item{graph}{An object of type \code{igraph}.}
}
\value{
igraph object.
}
\description{
Set likelihood attribute to vertices of graph using pval attribute.
}
